refactor(screen): split WindowCandidate and TextFragment into own files
Addresses PR review: one primary type per file (CLAUDE.md).
